Refactorisation du Front

+++ src/main/java/fr/isika/cda21/annuaire/models/Noeud.java
@@ -69,7 +69,7 @@ public class Noeud {
                 Noeud noeudDeLaListeChainee = this;
 
                 while (noeudDeLaListeChainee.listeChainee != LISTE_VIDE) { //Déplacement dans la liste chainée poour arriver à la fin
-                    raf.seek(noeudDeLaListeChainee.listeChainee * TAILLE_NOEUD_OCTETS); //positionne le curseur sur le prochain stagiaire de la liste chaine.
+                    raf.seek((long) noeudDeLaListeChainee.listeChainee * TAILLE_NOEUD_OCTETS); //positionne le curseur sur le prochain stagiaire de la liste chaine.
                     noeudDeLaListeChainee = GestionFichiers.lectureNoeud();
                 }
 
@@ -91,7 +91,7 @@ public class Noeud {
                 raf.seek(raf.length());
                 new Noeud(stagiaireAAjouter).ecrireNoeudBinaire(raf);
             } else {
-                raf.seek(this.filsGauche * TAILLE_NOEUD_OCTETS);
+                raf.seek((long) this.filsGauche * TAILLE_NOEUD_OCTETS);
                 Noeud noeudFilsGauche = GestionFichiers.lectureNoeud();
                 noeudFilsGauche.ajouterNoeud(stagiaireAAjouter, raf);
             }
@@ -103,7 +103,7 @@ public class Noeud {
                 raf.seek(raf.length());
                 new Noeud(stagiaireAAjouter).ecrireNoeudBinaire(raf);
             } else {
-                raf.seek(this.filsDroit * TAILLE_NOEUD_OCTETS);
+                raf.seek((long) this.filsDroit * TAILLE_NOEUD_OCTETS);
                 Noeud noeudFilsDroit = GestionFichiers.lectureNoeud();
                 noeudFilsDroit.ajouterNoeud(stagiaireAAjouter, raf);
             }
@@ -116,7 +116,7 @@ public class Noeud {
         Noeud noeudDeLaListeChainee = this;
 
         if (this.filsGauche != -1) {
-            raf.seek(this.filsGauche * TAILLE_NOEUD_OCTETS); // on positionne le curseur au niveau du fils gauche
+            raf.seek((long) this.filsGauche * TAILLE_NOEUD_OCTETS); // on positionne le curseur au niveau du fils gauche
             Noeud noeudFilsGauche = GestionFichiers.lectureNoeud(); // on lit le noeud fils gauche
             noeudFilsGauche.ordreAlphabetique(listeDeStagiaire, raf);
         } //G
@@ -125,14 +125,14 @@ public class Noeud {
         //N
 
         while (noeudDeLaListeChainee.listeChainee != LISTE_VIDE) {
-            raf.seek(noeudDeLaListeChainee.listeChainee * TAILLE_NOEUD_OCTETS);
+            raf.seek((long) noeudDeLaListeChainee.listeChainee * TAILLE_NOEUD_OCTETS);
             noeudDeLaListeChainee = GestionFichiers.lectureNoeud();
 
             listeDeStagiaire.add(noeudDeLaListeChainee.stagiaire);
         } //N
 
         if (this.filsDroit != -1) {
-            raf.seek(this.filsDroit * TAILLE_NOEUD_OCTETS); // on place le curseur au niveau du fils droit
+            raf.seek((long) this.filsDroit * TAILLE_NOEUD_OCTETS); // on place le curseur au niveau du fils droit
             Noeud noeudFilsDroit = GestionFichiers.lectureNoeud(); // on lit le noeud fils droit
             noeudFilsDroit.ordreAlphabetique(listeDeStagiaire, raf);
         } //D
@@ -150,7 +150,7 @@ public class Noeud {
 
             // Si une liste chaînée existe dans le noeud actuel, on se déplace pour lire la liste
             if (this.listeChainee != LISTE_VIDE) {
-                raf.seek(this.listeChainee * TAILLE_NOEUD_OCTETS);
+                raf.seek((long) this.listeChainee * TAILLE_NOEUD_OCTETS);
                 // Déclare noeudSuivant pour pouvoir transmettre la méthode
                 Noeud noeudSuivant = GestionFichiers.lectureNoeud();
                 noeudSuivant.rechercheStagiaire(listeResultats, stagiaireRecherche, raf);
@@ -162,7 +162,7 @@ public class Noeud {
                 // Retour visuel Aucun stagiaire correspondant
                 System.out.println("Aucun stagiaire correspondant à gauche.");
             } else {
-                raf.seek(this.filsGauche * TAILLE_NOEUD_OCTETS);
+                raf.seek((long) this.filsGauche * TAILLE_NOEUD_OCTETS);
                 Noeud noeudFilsGauche = GestionFichiers.lectureNoeud();
                 noeudFilsGauche.rechercheStagiaire(listeResultats, stagiaireRecherche, raf);
             }
@@ -173,7 +173,7 @@ public class Noeud {
                 System.out.println("Aucun stagiaire correspondant.");
             } else {
 
-                raf.seek(this.filsDroit * TAILLE_NOEUD_OCTETS);
+                raf.seek((long) this.filsDroit * TAILLE_NOEUD_OCTETS);
                 Noeud noeudFilsDroit = GestionFichiers.lectureNoeud();
                 noeudFilsDroit.rechercheStagiaire(listeResultats, stagiaireRecherche, raf);
             }
@@ -188,13 +188,13 @@ public class Noeud {
         //Recherche du stagiaire
         if (this.stagiaire.getNom().compareToIgnoreCase(stagiaireASupprimer.getNom()) > 0) {
 
-            raf.seek(this.filsGauche * TAILLE_NOEUD_OCTETS);// pour positionner le curseur au début du fils gauche on
+            raf.seek((long) this.filsGauche * TAILLE_NOEUD_OCTETS);// pour positionner le curseur au début du fils gauche on
             // multiplie l'index du fils gauche par la taille d'un noeud
             Noeud noeudFilsGauche = GestionFichiers.lectureNoeud(); // on lit les valeurs du fils gauche où le curseur
             // s'est arrêté pour stocker les informations
 
             this.filsGauche = noeudFilsGauche.supprimerNoeud(stagiaireASupprimer, raf);
-            raf.seek(indexDuStagiaire * TAILLE_NOEUD_OCTETS + Stagiaire.TAILLE_STAGIAIRE_OCTETS + TAILLE_LISTECHAINEE_OCTETS); //car on est au début de notre Noeud
+            raf.seek((long) indexDuStagiaire * TAILLE_NOEUD_OCTETS + Stagiaire.TAILLE_STAGIAIRE_OCTETS + TAILLE_LISTECHAINEE_OCTETS); //car on est au début de notre Noeud
             raf.writeInt(this.filsGauche);
 
         } else if (this.stagiaire.getNom().compareToIgnoreCase(stagiaireASupprimer.getNom()) < 0) { // on ne veut pas le contraire de
@@ -202,11 +202,11 @@ public class Noeud {
             // strictement plus grand ou égal)
             // mais on veut juste le
             // strictement plus grand.
-            raf.seek(this.filsDroit * TAILLE_NOEUD_OCTETS);
+            raf.seek((long) this.filsDroit * TAILLE_NOEUD_OCTETS);
             Noeud noeudFilsDroit = GestionFichiers.lectureNoeud();
 
             this.filsDroit = noeudFilsDroit.supprimerNoeud(stagiaireASupprimer, raf);
-            raf.seek(indexDuStagiaire * TAILLE_NOEUD_OCTETS + Stagiaire.TAILLE_STAGIAIRE_OCTETS + TAILLE_LISTECHAINEE_OCTETS + TAILLE_FILS_GAUCHE_OCTETS); //car on est au début de notre Noeud
+            raf.seek((long) indexDuStagiaire * TAILLE_NOEUD_OCTETS + Stagiaire.TAILLE_STAGIAIRE_OCTETS + TAILLE_LISTECHAINEE_OCTETS + TAILLE_FILS_GAUCHE_OCTETS); //car on est au début de notre Noeud
             raf.writeInt(this.filsDroit);
 
         } else { // cas où on a trouvé le nom du stagiaire à supprimer
@@ -226,16 +226,16 @@ public class Noeud {
                     // curseur.
 
                     Noeud noeudDeRemplacement = this.getSuccesseur(raf);
-                    raf.seek(indexDuStagiaire * TAILLE_NOEUD_OCTETS); // on a positionné notre curseur au bon
+                    raf.seek((long) indexDuStagiaire * TAILLE_NOEUD_OCTETS); // on a positionné notre curseur au bon
                     // endroit: le début du stagiaire à supprimer.
                     noeudDeRemplacement.stagiaire.ecritureStagiaireBinaire(raf);// on a écrit les informations de notre
                     // stagiaire successeur et de sa liste chaînée dans notre noeud
                     raf.writeInt(noeudDeRemplacement.listeChainee);
 
-                    raf.seek(this.filsDroit * TAILLE_NOEUD_OCTETS);
+                    raf.seek((long) this.filsDroit * TAILLE_NOEUD_OCTETS);
                     Noeud noeudFilsDroit = GestionFichiers.lectureNoeud();
                     this.filsDroit = noeudFilsDroit.supprimerNoeud(noeudDeRemplacement.stagiaire, raf);
-                    raf.seek(indexDuStagiaire * TAILLE_NOEUD_OCTETS + Stagiaire.TAILLE_STAGIAIRE_OCTETS + TAILLE_LISTECHAINEE_OCTETS + TAILLE_FILS_GAUCHE_OCTETS); //car on est au début de notre Noeud
+                    raf.seek((long) indexDuStagiaire * TAILLE_NOEUD_OCTETS + Stagiaire.TAILLE_STAGIAIRE_OCTETS + TAILLE_LISTECHAINEE_OCTETS + TAILLE_FILS_GAUCHE_OCTETS); //car on est au début de notre Noeud
                     raf.writeInt(this.filsDroit);
 
                     return indexDuStagiaire;
@@ -246,13 +246,13 @@ public class Noeud {
                 Noeud noeudActuel = this;
                 int indexNoeudActuel = indexDuStagiaire;
 
-                raf.seek(noeudActuel.listeChainee  * TAILLE_NOEUD_OCTETS);
+                raf.seek((long) noeudActuel.listeChainee * TAILLE_NOEUD_OCTETS);
                 Noeud noeudChaineSuivant = GestionFichiers.lectureNoeud();
 
-                raf.seek(indexNoeudActuel * TAILLE_NOEUD_OCTETS);
+                raf.seek((long) indexNoeudActuel * TAILLE_NOEUD_OCTETS);
                 noeudChaineSuivant.stagiaire.ecritureStagiaireBinaire(raf);
 
-                raf.seek(indexDuStagiaire * TAILLE_NOEUD_OCTETS + Stagiaire.TAILLE_STAGIAIRE_OCTETS);
+                raf.seek((long) indexDuStagiaire * TAILLE_NOEUD_OCTETS + Stagiaire.TAILLE_STAGIAIRE_OCTETS);
                 raf.writeInt(noeudChaineSuivant.listeChainee);
 
             } else  { //cas où le stagiaire à supprimer est dans la liste chainée mais pas au début.
@@ -260,13 +260,13 @@ public class Noeud {
                 int indexNoeudCourant = indexDuStagiaire;
                 int indexPrecedent = LISTE_VIDE;
                 while (noeudCourant.stagiaire.compareTo(stagiaireASupprimer) != 0) {
-                    raf.seek(noeudCourant.listeChainee * TAILLE_NOEUD_OCTETS);
+                    raf.seek((long) noeudCourant.listeChainee * TAILLE_NOEUD_OCTETS);
                     indexPrecedent = indexNoeudCourant;
                     indexNoeudCourant = noeudCourant.listeChainee;
                     noeudCourant = GestionFichiers.lectureNoeud();
                 }
 
-                raf.seek(indexPrecedent * TAILLE_NOEUD_OCTETS + Stagiaire.TAILLE_STAGIAIRE_OCTETS);
+                raf.seek((long) indexPrecedent * TAILLE_NOEUD_OCTETS + Stagiaire.TAILLE_STAGIAIRE_OCTETS);
                 raf.writeInt(noeudCourant.listeChainee);
             }
         }
@@ -276,10 +276,10 @@ public class Noeud {
 
     //pour récupérer les informations du noeud qui succecede celui à supprimer (ici le plus petit des plus grands)
     private Noeud getSuccesseur(RandomAccessFile raf) throws IOException {
-        raf.seek(this.filsDroit * TAILLE_NOEUD_OCTETS);
+        raf.seek((long) this.filsDroit * TAILLE_NOEUD_OCTETS);
         Noeud noeudTemporaire = GestionFichiers.lectureNoeud();
         while (noeudTemporaire.filsGauche != FILS_NUL) {
-            raf.seek(noeudTemporaire.filsGauche * TAILLE_NOEUD_OCTETS);
+            raf.seek((long) noeudTemporaire.filsGauche * TAILLE_NOEUD_OCTETS);
             noeudTemporaire = GestionFichiers.lectureNoeud();
         }
         return noeudTemporaire;
